feat(player_lifetime): split playerlifetime control as a module from player
