+++ src/player/mod.rs
@@ -12,36 +12,17 @@ mod config;
 mod component;
 mod event;
 
-use self::event::PlayerSpawnEvent;
 use self::spawn::spawn_player;
-use self::control::player_input_system;
+
+use crate::event::RequestControl;
+use crate::event::PlayerSpawnEvent;
 
 pub struct PlayerPlugin;
 
 impl Plugin for PlayerPlugin {
     fn build(&self, app: &mut App) {
         app.add_event::<PlayerSpawnEvent>()
-        .add_systems(Startup, queue_for_player_setup_event)
+        .add_event::<RequestControl>()
         .add_systems(Update, spawn_player)
-        .add_systems(Update, player_input_system);
     }
-}
-
-use self::component::ControlScheme;
-
-fn queue_for_player_setup_event(
-    asset_server: Res<AssetServer>,
-    mut events: EventWriter::<PlayerSpawnEvent>,
-) {
-    let tex: Handle<Image> = asset_server.load("portrait_rainey.png");
-    events.write(PlayerSpawnEvent {
-        texture: tex,
-        position: Vec2::new(-500.0,-200.0),
-        controls: ControlScheme{
-            up: KeyCode::KeyW,
-            left: KeyCode::KeyA,
-            right: KeyCode::KeyD,
-        },
-        mass: Some(50.0),
-    });
 }
